// src/app/modules/storeFront/models/paginationResponseOfOrdersDto.ts
/**
 * .NET 6.0 WebAPI - Clean Architecture
 * Clean Architecture Template for .NET 6.0 WebApi built with Multitenancy Support.
 *
 * OpenAPI spec version: v4
 * Contact: hello@codewithmukesh.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OrdersDto } from './ordersDto';

export interface PaginationResponseOfOrdersDto { 
    data?: Array<OrdersDto>;
    currentPage?: number;
    totalPages?: number;
    totalCount?: number;
    pageSize?: number;
    hasPreviousPage?: boolean;
    hasNextPage?: boolean;
}