// src/app/modules/storeFront/models/paginationResponseOfProductDto.ts
/**
 * .NET 6.0 WebAPI - Clean Architecture
 * Clean Architecture Template for .NET 6.0 WebApi built with Multitenancy Support.
 *
 * OpenAPI spec version: v2
 * Contact: hello@codewithmukesh.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProductDto } from './productDto';

export interface PaginationResponseOfProductDto { 
    data?: Array<ProductDto>;
    currentPage?: number;
    totalPages?: number;
    totalCount?: number;
    pageSize?: number;
    hasPreviousPage?: boolean;
    hasNextPage?: boolean;
}